Integrated the reviews: refactor the tests in the existing test classes and added new tests eliminating the proctime assignment modify the condition for the SQL rule
