+++ flink-libraries/flink-table/src/test/scala/org/apache/flink/table/api/scala/stream/sql/WindowAggregateTest.scala
@@ -24,6 +24,19 @@ import org.apache.flink.table.plan.logical.{EventTimeTumblingGroupWindow, Proces
 import org.apache.flink.table.utils.{StreamTableTestUtil, TableTestBase}
 import org.apache.flink.table.utils.TableTestUtil._
 import org.junit.Test
+import org.apache.flink.table.api.TableEnvironment
+import org.apache.flink.table.api.scala.stream.utils.{
+  StreamingWithStateTestBase,
+  StreamITCase,
+  StreamTestData
+}
+import org.apache.flink.types.Row
+import scala.collection.mutable
+import org.apache.flink.streaming.api.functions.AssignerWithPunctuatedWatermarks
+import org.apache.flink.streaming.api.watermark.Watermark
+import org.junit.Assert._
+import org.apache.flink.streaming.api.scala.StreamExecutionEnvironment
+
 
 class WindowAggregateTest extends TableTestBase {
   private val streamUtil: StreamTableTestUtil = streamTestUtil()
@@ -239,4 +252,179 @@ class WindowAggregateTest extends TableTestBase {
       )
     streamUtil.verifySql(sql, expected)
   }
+  
+   @Test
+  def testMaxAggregatation(): Unit = {
+
+    val env = StreamExecutionEnvironment.getExecutionEnvironment
+    val tEnv = TableEnvironment.getTableEnvironment(env)
+    env.setParallelism(1)
+    StreamITCase.testResults = mutable.MutableList()
+
+    val sqlQuery = "SELECT a, MAX(c) OVER (PARTITION BY a ORDER BY procTime()" +
+      "RANGE BETWEEN INTERVAL '10' SECOND PRECEDING AND CURRENT ROW) AS maxC FROM MyTable"
+
+    val t = StreamTestData.get5TupleDataStream(env)
+        .assignTimestampsAndWatermarks(new ProcTimeTimestamp())
+        .toTable(tEnv).as('a, 'b, 'c, 'd, 'e)
+    tEnv.registerTable("MyTable", t)
+
+    val result = tEnv.sql(sqlQuery).toDataStream[Row]
+    result.addSink(new StreamITCase.StringSink)
+    env.execute()
+
+    val expected = mutable.MutableList(
+      "1,0",
+      "2,1",
+      "2,2",
+      "3,3",
+      "3,4",
+      "3,5",
+      "4,6",
+      "4,7",
+      "4,8",
+      "4,9",
+      "5,10",
+      "5,11",
+      "5,12",
+      "5,13",
+      "5,14")
+    assertEquals(expected.sorted, StreamITCase.testResults.sorted)
+  }
+
+  @Test
+  def testMinAggregatation(): Unit = {
+
+    val env = StreamExecutionEnvironment.getExecutionEnvironment
+    val tEnv = TableEnvironment.getTableEnvironment(env)
+    env.setParallelism(1)
+    StreamITCase.testResults = mutable.MutableList()
+
+    val sqlQuery = "SELECT a, MIN(c) OVER (PARTITION BY a ORDER BY procTime()" +
+      "RANGE BETWEEN INTERVAL '10' SECOND PRECEDING AND CURRENT ROW) AS minC FROM MyTable"
+
+    val t = StreamTestData.get5TupleDataStream(env)
+        .assignTimestampsAndWatermarks(new ProcTimeTimestamp())
+        .toTable(tEnv).as('a, 'b, 'c, 'd, 'e)
+    tEnv.registerTable("MyTable", t)
+
+    val result = tEnv.sql(sqlQuery).toDataStream[Row]
+    result.addSink(new StreamITCase.StringSink)
+    env.execute()
+
+    val expected = mutable.MutableList(
+      "1,0",
+      "2,1",
+      "2,1",
+      "3,3",
+      "3,3",
+      "3,3",
+      "4,6",
+      "4,6",
+      "4,6",
+      "4,6",
+      "5,10",
+      "5,10",
+      "5,10",
+      "5,10",
+      "5,10")
+    assertEquals(expected.sorted, StreamITCase.testResults.sorted)
+  }
+
+  @Test
+  def testSumAggregatation(): Unit = {
+
+    val env = StreamExecutionEnvironment.getExecutionEnvironment
+    val tEnv = TableEnvironment.getTableEnvironment(env)
+    env.setParallelism(1)
+    StreamITCase.testResults = mutable.MutableList()
+
+    val sqlQuery = "SELECT a, SUM(c) OVER (PARTITION BY a ORDER BY procTime()" +
+      "RANGE BETWEEN INTERVAL '10' SECOND PRECEDING AND CURRENT ROW) AS sumC FROM MyTable"
+
+    val t = StreamTestData.get5TupleDataStream(env)
+        .assignTimestampsAndWatermarks(new ProcTimeTimestamp())
+        .toTable(tEnv).as('a, 'b, 'c, 'd, 'e)
+    tEnv.registerTable("MyTable", t)
+
+    val result = tEnv.sql(sqlQuery).toDataStream[Row]
+    result.addSink(new StreamITCase.StringSink)
+    env.execute()
+
+    val expected = mutable.MutableList(
+      "1,0",
+      "2,1",
+      "2,3",
+      "3,12",
+      "3,3",
+      "3,7",
+      "4,13",
+      "4,21",
+      "4,30",
+      "4,6",
+      "5,10",
+      "5,21",
+      "5,33",
+      "5,46",
+      "5,60")
+    assertEquals(expected.sorted, StreamITCase.testResults.sorted)
+  }
+
+  @Test
+  def testAvgAggregatation(): Unit = {
+
+    val env = StreamExecutionEnvironment.getExecutionEnvironment
+    val tEnv = TableEnvironment.getTableEnvironment(env)
+    env.setParallelism(1)
+    StreamITCase.testResults = mutable.MutableList()
+
+    val sqlQuery = "SELECT a, AVG(c) OVER (PARTITION BY a ORDER BY procTime()" +
+      "RANGE BETWEEN INTERVAL '10' SECOND PRECEDING AND CURRENT ROW) AS avgC FROM MyTable"
+
+    val t = StreamTestData.get5TupleDataStream(env)
+         .assignTimestampsAndWatermarks(new ProcTimeTimestamp())
+         .toTable(tEnv).as('a, 'b, 'c, 'd, 'e)
+    
+    tEnv.registerTable("MyTable", t)
+
+    val result = tEnv.sql(sqlQuery).toDataStream[Row]
+    result.addSink(new StreamITCase.StringSink)
+    env.execute()
+
+    val expected = mutable.MutableList(
+      "1,0",
+      "2,1",
+      "2,1",
+      "3,3",
+      "3,3",
+      "3,4",
+      "4,6",
+      "4,6",
+      "4,7",
+      "4,7",
+      "5,10",
+      "5,10",
+      "5,11",
+      "5,11",
+      "5,12")
+    assertEquals(expected.sorted, StreamITCase.testResults.sorted)
+  }
+  
 }
+
+class ProcTimeTimestamp
+      extends AssignerWithPunctuatedWatermarks[(Int, Long, Int, String, Long)] {
+
+    override def checkAndGetNextWatermark(
+      lastElement: (Int, Long, Int, String, Long),
+      extractedTimestamp: Long): Watermark = {
+      null
+    }
+
+    override def extractTimestamp(
+      element: (Int, Long, Int, String, Long),
+      previousElementTimestamp: Long): Long = {
+      System.currentTimeMillis()
+    }
+  }
+
